relocate tests for orderliness
